// src/api/model/responseResetDto.ts
/**
 * Service Auth
 * API dashboard
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ResetDto } from './resetDto';


export interface ResponseResetDto {
    reset: Array<ResetDto>;
    total: number;
}