// src/api/model/resetRequestDto.ts
/**
 * Service Auth
 * API dashboard
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ResetRequestDto {
    limit: number;
    page: number;
    from: number;
    to: number;
    user: string;
    tipo: string;
    estado: string;
}